# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: SuperNØva Smart Locker API
  description: |
    API RESTful per il sistema di smart locker SuperNØva.

    ## Autenticazione
    L'API utilizza JWT (JSON Web Tokens) per l'autenticazione. Includi il token nell'header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - 100 richieste per 15 minuti per IP
    - Risposta 429 quando il limite viene superato

    ## Errori
    Tutte le risposte di errore seguono il formato:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Descrizione dell'errore"
      }
    }
    ```
  version: 1.0.0
  contact:
    name: SuperNØva Support
    email: api@supernova-locker.it
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.supernova-locker.it/v1
    description: Production server
...
paths:
...
  /access/{deviceId}/permissions:
    get:
      tags: [Access Control]
      summary: Lista permessi dispositivo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Lista permessi
  /access/{deviceId}/share:
    post:
      tags: [Access Control]
      summary: Condividi accesso
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [viewer, manager]
                expires_at:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Accesso condiviso
  /access/{deviceId}/share/{permissionId}:
    delete:
      tags: [Access Control]
      summary: Revoca accesso condiviso
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
        - name: permissionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Accesso revocato
  /access/{deviceId}/temp-code:
    post:
      tags: [Access Control]
      summary: Genera codice temporaneo
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [validity_minutes]
              properties:
                validity_minutes:
                  type: integer
                  minimum: 5
                  maximum: 1440
                  example: 30
                note:
                  type: string
                  example: Consegna DHL
      responses:
        '201':
          description: Codice generato
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      code_id:
                        type: string
                      code:
                        type: string
                        example: A1B2C3
                      expires_at:
                        type: string
                        format: date-time
  /access/{deviceId}/verify-code:
    post:
      tags: [Access Control]
      summary: Verifica e usa codice temporaneo
      parameters:
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  example: A1B2C3
      responses:
        '200':
          description: Dispositivo sbloccato
        '401':
          description: Codice non valido o scaduto

  # ============================================
  # Logs Endpoints
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    deviceId:
      name: deviceId
      in: path
      required: true
      description: ID univoco del dispositivo
      schema:
        type: string
        example: dev_cassapanca_001
...